fix(script): Move shared scope owner
