+++ src/scripting/script_runtime.cpp
@@ -11,8 +11,12 @@
 
 #include <algorithm>
 #include <deque>
+#include <iomanip>
 #include <mutex>
+#include <sstream>
+#include <type_traits>
 #include <unordered_map>
+#include <vector>
 
 namespace scripting {
   namespace {
@@ -87,6 +91,82 @@ namespace scripting {
           break;
         }
       }
+    }
+
+    std::string escapeRuntimeKeyToken(std::string_view token) {
+      std::string out;
+      out.reserve(token.size());
+      for (char ch : token) {
+        if (ch == '\\' || ch == '|' || ch == '=' || ch == ',' || ch == ':') {
+          out.push_back('\\');
+        }
+        out.push_back(ch);
+      }
+      return out;
+    }
+
+    std::string encodeRuntimeSettingValue(const WidgetSettingValue& value) {
+      return std::visit(
+          [](const auto& concrete) -> std::string {
+            using T = std::decay_t<decltype(concrete)>;
+            if constexpr (std::is_same_v<T, bool>) {
+              return std::string("b:") + (concrete ? "1" : "0");
+            } else if constexpr (std::is_same_v<T, std::int64_t>) {
+              return std::string("i:") + std::to_string(concrete);
+            } else if constexpr (std::is_same_v<T, double>) {
+              std::ostringstream out;
+              out << "d:" << std::setprecision(17) << concrete;
+              return out.str();
+            } else if constexpr (std::is_same_v<T, std::string>) {
+              return std::string("s:") + escapeRuntimeKeyToken(concrete);
+            } else if constexpr (std::is_same_v<T, std::vector<std::string>>) {
+              std::ostringstream out;
+              out << "v:" << concrete.size() << ':';
+              for (std::size_t i = 0; i < concrete.size(); ++i) {
+                if (i != 0) {
+                  out << ',';
+                }
+                out << escapeRuntimeKeyToken(concrete[i]);
+              }
+              return out.str();
+            } else {
+              return std::string{};
+            }
+          },
+          value
+      );
+    }
+
+    std::string buildSharedScriptRuntimeKey(
+        std::string_view baseKey, std::string_view scriptPath, const ScriptWidgetSettings& settings
+    ) {
+      std::vector<std::string> settingKeys;
+      settingKeys.reserve(settings.size());
+      for (const auto& [key, value] : settings) {
+        (void)value;
+        settingKeys.push_back(key);
+      }
+      std::sort(settingKeys.begin(), settingKeys.end());
+
+      std::ostringstream out;
+      out
+          << "base="
+          << escapeRuntimeKeyToken(baseKey)
+          << "|script="
+          << escapeRuntimeKeyToken(scriptPath)
+          << "|settings=";
+      for (std::size_t i = 0; i < settingKeys.size(); ++i) {
+        if (i != 0) {
+          out << '|';
+        }
+        const auto& key = settingKeys[i];
+        const auto it = settings.find(key);
+        if (it == settings.end()) {
+          continue;
+        }
+        out << escapeRuntimeKeyToken(key) << '=' << encodeRuntimeSettingValue(it->second);
+      }
+      return out.str();
     }
   } // namespace
 
@@ -631,16 +711,20 @@ namespace scripting {
   }
 
   SharedScriptRuntimeAcquireResult SharedScriptRuntimeRegistry::acquire(
-      const std::string& key, ScriptWidgetSettings settings, ScriptApiContext& api, ClipboardService* clipboard
+      std::string_view baseKey, std::string_view scriptPath, ScriptWidgetSettings settings, ScriptApiContext& api,
+      ClipboardService* clipboard
   ) {
     static std::mutex mutex;
     static std::unordered_map<std::string, std::weak_ptr<ScriptRuntime>> runtimes;
+
+    const std::string key = buildSharedScriptRuntimeKey(baseKey, scriptPath, settings);
 
     std::lock_guard lock(mutex);
     if (auto it = runtimes.find(key); it != runtimes.end()) {
       if (auto runtime = it->second.lock()) {
         return {.runtime = std::move(runtime), .created = false};
       }
+      runtimes.erase(it);
     }
 
     auto runtime = std::make_shared<ScriptRuntime>(key, std::move(settings), api, clipboard);
